backend/requirements.txt
# FastAPI Backend Dependencies for Digital Cassette

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Serialization
pydantic==2.6.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-auth==2.27.0

# HTTP Client
httpx==0.26.0
requests==2.31.0
